+++ src/eip4844/api.c
@@ -16,7 +16,11 @@
 
 #include "api.h"
 #include "common/alloc.h"
+#include "common/common.h"
 #include "common/fr.h"
+#include "common/g1.h"
+#include "common/ret.h"
+#include "common/settings.h"
 
 #include <assert.h> /* For assert */
 #include <stdlib.h> /* For NULL */
