+++ src/setup/api.c
@@ -16,7 +16,6 @@
 
 #include "api.h"
 #include "common/alloc.h"
-#include "common/fr.h"
 #include "eip7594/api.h"
 #include "eip7594/fft.h"
